requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.6
django-bootstrap3==8.2.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
gunicorn==19.7.0
markdown2==2.3.3
pytz==2016.10
static3==0.7.0
psycopg2>=2.6.1